+++ keepalived/check/check_daemon.c
@@ -5,7 +5,7 @@
  *
  * Part:        Healthcheckrs child process handling.
  *
- * Version:     $Id: check_daemon.c,v 1.1.2 2003/09/08 01:18:41 acassen Exp $
+ * Version:     $Id: check_daemon.c,v 1.1.3 2003/09/29 02:37:13 acassen Exp $
  *
  * Author:      Alexandre Cassen, <acassen@linux-vs.org>
  *
@@ -229,7 +229,7 @@ start_check_child(void)
 			       , strerror(errno));
 		return -1;
 	} else if (pid) {
-		int poll_delay = (wdog_delay_check) ? wdog_delay_check : WATCHDOG_DELAY;
+		long poll_delay = (wdog_delay_check) ? wdog_delay_check : WATCHDOG_DELAY;
 		checkers_child = pid;
 		syslog(LOG_INFO, "Starting Healthcheck child process, pid=%d"
 			       , pid);
