+++ keepalived/vrrp/vrrp_daemon.c
@@ -5,7 +5,7 @@
  *
  * Part:        VRRP child process handling.
  *
- * Version:     $Id: vrrp_daemon.c,v 1.1.2 2003/09/08 01:18:41 acassen Exp $
+ * Version:     $Id: vrrp_daemon.c,v 1.1.3 2003/09/29 02:37:13 acassen Exp $
  *
  * Author:      Alexandre Cassen, <acassen@linux-vs.org>
  *
@@ -235,7 +235,7 @@ start_vrrp_child(void)
 			       , strerror(errno));
 		return -1;
 	} else if (pid) {
-		int poll_delay = (wdog_delay_vrrp) ? wdog_delay_vrrp : WATCHDOG_DELAY;
+		long poll_delay = (wdog_delay_vrrp) ? wdog_delay_vrrp : WATCHDOG_DELAY;
 		vrrp_child = pid;
 		syslog(LOG_INFO, "Starting VRRP child process, pid=%d"
 			       , pid);
